Make starting the measurements more reliable
The flow sensors only need to be put into measurement mode once, afterwards
a simple i2c read is enough to read out the current values. The driver now
remembers if the measurement has been started already. If a measurement
fails, the command is resent on the next try.

Also the sf05 needs quite a while for the first measurements to be available,
so we wait after triggering it.

Change-Id: I4c76b50f776b3a17d5816a85aac3616da7d9f618